Prevent headless exception when deleting a node via the API
